+++ s/flatview/flatview.ts
@@ -1,15 +1,21 @@
 
 import {CSSResultArray} from "lit"
+import {Flat} from "../flatstate/flat.js"
 import {make_view_directive} from "./parts/directive.js"
 import {FlatviewOptions, FlatviewRenderer, FlatviewSetup} from "./parts/types.js"
 
-export {Flatview, FlatviewOptions, FlatviewSetup, FlatviewContext, FlatviewRenderer} from "./parts/types.js"
+export * from "./parts/types.js"
 
-export const flatview = (options: FlatviewOptions = {}) => ({
+export const flatview = ({
+		flat = new Flat(),
+		strict = true,
+	}: FlatviewOptions = {}) => ({
 
 	state: <S extends {}>(initstate: S = {} as any) => ({
 
-		actions: <A extends {}>(initactions: (state: S) => A = () => ({} as any)) => ({
+		actions: <A extends {}>(
+				initactions: (state: S) => A = () => ({} as any)
+			) => ({
 
 			setup: (setup: FlatviewSetup<S, A> = () => () => {}) => ({
 
@@ -18,7 +24,8 @@ export const flatview = (options: FlatviewOptions = {}) => ({
 					css: (...css: CSSResultArray) => (
 
 						make_view_directive<S, A, P>({
-							...options,
+							flat,
+							strict,
 							initstate,
 							initactions,
 							setup,